refactor(cli): split generate plan-building into generate_plans.py (#1331)
Move the plan-building half of the generate service — the enum/format
maps, GenerationKind, GenerationPlan, GenerationPlanValidationError,
build_generation_plan, _common, the per-kind builders, and _BUILDERS —
verbatim into a new cli/services/generate_plans.py. GenerationKind moves
with the plan types (it is consumed by GenerationPlan.kind) so the new
module imports nothing from generate.py, avoiding a cycle.

generate.py keeps the executor (GenerationExecutionResult, _KIND_TO_METHOD,
_build_call_kwargs, execute_generation) and re-exports GenerationKind,
GenerationPlan, GenerationPlanValidationError, build_generation_plan, and
the private _INFOGRAPHIC_STYLE_MAP (imported directly by generate_cmd.py)
so every external importer of notebooklm.cli.services.generate is
unaffected. __all__ is unchanged.

generate.py drops from 926 to 257 lines (off the size ratchet); the new
generate_plans.py is 707. Bodies moved verbatim — no behavior change.
